### 2020-05-27 #380 @mcguinlu
medrxivr: Accessing and searching medRxiv preprint data in R
Submitting Author: Luke McGuinness (@mcguinlu)
Repository: https://github.com/mcguinlu/medrxivr
Version submitted: 0.0.2
Version accepted: TBD

---

```
Package: medrxivr
Title: Access MedRxiv Preprint Data
Version: 0.0.2
    person("Luke", "McGuinness",
           email = "luke.mcguinness@bristol.ac.uk",
    person("Lena", "Schmidt",
           role = "aut",
           comment = c(ORCID = "0000-0003-0709-8226")))
Description: The medRxiv <https://www.medrxiv.org/> repository is a free online
    archive and distribution server for complete but unpublished manuscripts
    (preprints) in the medical, clinical, and related health sciences. medrxivr
    provides programmatic access to both medRxiv API <https://api.biorxiv.org/>
    and a static snapshot of database, which is updated daily. Users can then
    search for relevant records using regular expressions and Boolean logic, and
    can easily download the full-text PDFs of preprints matching their search
    criteria.
    rvest,
    xml2,
    curl,
    httr,
    rlang
    covr,
    kableExtra

```

	- [ ] workflow automataion

medrxivr allows users to programmatically access data from [medRxiv](https://www.medrxiv.org/), a preprint respository for papers in medical, clinical, and related health sciences. The package also allows user to readily perform and document reproducible literature searches of the medRxiv database.

The primary target of this package is systematic reviewers (i.e. me!), who frequently wish both to use more complicated queries (e.g. regular expresssions/Boolean combinations) when searching medRxiv than the official site currrently allows for, and who also wish to be easily able to download the full text PDFs of records matching their search. `medrxivr` helps with both of these challenges. However, anyone who wishes to find and retrieve relevant medRxiv records in R, for example to explore the distribution of preprints by subject area, will find the package useful.

As far as I am aware, no other package allows users to access medRxiv data in R.

Issue: https://github.com/ropensci/software-review/issues/369

    - [X] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.3860024

Tagging my co-author @L-ENA for reference.
