+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   building-complex-search-strategies: building-complex-search-strategies.html
   medrxiv-api: medrxiv-api.html
   medrxivr: medrxivr.html
-last_built: 2020-05-27T11:09Z
+last_built: 2020-05-27T14:20Z
 
