### 2020-06-04 ropensci/medrxivr#4 @mcguinlu
Fix spelling
Requested by editor:

https://github.com/ropensci/software-review/issues/380
