### 2020-06-04 ropensci/medrxivr#5 @mcguinlu
Run styler
Requested by editor:

https://github.com/ropensci/software-review/issues/380
