### 2020-06-04 #380 @mcguinlu
Hi @maurolepore

Thanks for your inital review of our package. I've gone through it and try to address each point below:

**ml1: Overlap with fulltext**
Personally, I think that fulltext and medrxivr should continue to be two seperate packages, but that there is the potential for intergration between the two (as occured with the `aRxiv` package). In all honesty, I completely forgot to reply to the issue on `fulltext` 🤦‍♂️- sorry @sckott!  The reason I think they should be seperate is two fold:

- In the first instance, the two packages take fundamentally different approaches to searching. At present, the `medrxivr` workflow is to create a local copy of the whole medRxiv repository via the API (or maintained static snapshot), and then search it locally using the `mx_search()` function to find relevant articles [i.e. all data -> local search -> results]. In comparison, `fulltext` takes the same approach to search bioRxiv as the `biorxivr` package does (**note:** medRxiv and bioRxiv are very similar, so I am using `fulltext`'s approach to bioRxiv for comparison here). They both paste the search (see [this code line](https://github.com/ropensci/fulltext/blob/d13b0b40f170bcda30cd538881f7a7b7183fa41e/R/bio_utils.R#L23)) on after the base URL (`https://www.biorxiv.org/search`) and scrape the resulting page(s), essentially mimicing what would happen if you performed a search on the site itself [i.e.  remote search -> results]. The search functionality offered by this approach is dependent on what the site itself offers, and so therefore is not as comprehensive as that offered by `medrxivr` (e.g. you can't use regexes to define capitalisation/alternative spellings, or use the NEAR operator). In addition, using the search functionality of the site itself (i.e. by pasting the query onto the `search/` path) [has been shown to not be very reproducible/transparent](https://twitter.com/mcguinlu/status/1257394058743427072), which is what originally motivated the development of `medrxivr`.

- Secondly, as far as I can tell (correct me if I am wrong @sckott), search strings for `fulltext` do not vary based on the database searched - for example, if you use `ft_search()` without specifying the source, it uses the same string to search every database. This would cause issues for anything beyond a simple search, as `medrxivr` allows for advanced search strings that would not be compatible with other data sources. Based on this, my argument is that `medrxivr` should be a standalone full package, and a simple restricted version of the `medrxivr` search could be implemented in `fulltext`, provided @sckott is happy to implement the `medrxivr` workflow [i.e. download data then search] within `fulltext`.

**ml2: Overlap with biorxivr**
Thanks for highlighting this. I did come across this package while developing `medrxivr` - however, the last work on this package took place 5 years ago, before introduction of the API which you refer to in your query. In addition, while the base URL of the API contains "bioRxiv" (e.g. `https://api.biorxiv.org/`), this is only because the same organisation (Cold Spring Harbour) is responsible for both repositories. The actually endpoint for the medRxiv API is `https://api.biorxiv.org/details/medrxiv/[interval]/[cursor]/[format]`. Finally, as mentioned above, `biorxivr` relies on the search functionality offered by the site (e.g. by pasting the query on after the `search/` path) rather than performing the searches itself.

**ml3: Spelling**
I've added this as an issue (https://github.com/mcguinlu/medrxivr/issues/4) and plan to address it soon.

**ml4/ml5: Test coverage**
In hindsight, I should have highlighted this as a potential sticking point in my inital submission. The single file which is dragging down the average coverage contains `mx_download()`, which takes the dataframe of records identified by the user as relevant and downloads a PDF for each one. In fact, this function has a test suite (see https://github.com/mcguinlu/medrxivr/blob/master/tests/testthat/test-download.R), but because it manipulates files and folders on a users machine, I have these tests set to skip_on_CRAN, which in turn means that `covr` doesn't pick them up. This is something I was hoping to get feedback on during the review process, as I am not sure if this is best practice or what a viable alternative would be?

**ml6: styler**
I've added this as an issue (https://github.com/mcguinlu/medrxivr/issues/5) and plan to address it soon.

Hopefully this addresses your inital concerns, but please do let me know if anything is unclear, if my responses are insufficient, or if you need further details!
