### 2020-06-06 #380 @maurolepore
Thanks @mcguinlu! I think {medrxivr} merits to move to the next stage. I'll now start searching for reviewers.

### ml1 and ml2

Here is my conclusion. I base it on your answers above, and on this quote from rOpenSci's
guidelines on [overlap](https://devguide.ropensci.org/policies.html#overlap):

> "An R package that replicates the functionality of an existing R package
may be considered for inclusion in the rOpenSci suite if it significantly
improves on alternatives in any repository (RO, CRAN, BioC) by being ...
better in usability and performance".

I considered the packages {medrxivr}, {fulltext} and {biorxivr}. I see an
overlap in _what_ the packages aim to do but not in _how_ they do
it. This would not justify the overlap in general; but in this case I think it does.

Compared to the other packages, {medrxivr} searches locally. This ensures the
results can be reproduced; and enables searching with regular expressions. The
different approach to searching also means that to integrate {medrxivr} into
the other packages seems challenging. This might be eventually possible,
but first {medrxivr} may need to mature independently.

### ml3 to ml6

@mcguinlu, please let me know or check the boxes as you address these issues.

### ml7

@mcguinlu, I see the positive aspects of the "local" approach to searching that {medrxivr} implements; but I understand that {medrxivr} downloads the entire database. I worry this may not scale up. Here are some questions I have; you may discuss them directly with the reviewers:

* How big is the database? How fast does it grow?  And how long does it take to download it in a range of reasonable conditions? What happens in a range of extreme conditions?

* Is the process transparent and "polite" to the user?
