### 2020-06-06 #380 @maurolepore
@mcguinlu, please do this (from [these guidelines](https://devguide.ropensci.org/editorguide.html)):
