### 2020-06-06 #380 @maurolepore
(ml8) @mcguinlu have you suggested any reviewer?

(The editor guidelines suggest you might but I fail to find them here or at https://github.com/ropensci/software-review/issues/369.)
