### 2020-06-28 #380 @tts
Hi @maurolepore  and @mcguinlu - here is my review. Thanks for this opportunity, and all the best for the package!

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

[medRxiv](https://www.medrxiv.org/) has been accepting preprints [for a year now](https://twitter.com/mcguinlu/status/1263386069153783809). Their API does not offer any search capabilities, so clearly `medrxivr` has a function to fill. Regular expressions and Boolean logic are state-of-the-art ways to fine-tune search queries, so users of this package should be happy. In addition, you can download all searched and found preprints as PDF files, which is handy and helpful.

Although the target group and the goal of the package are clearly defined, it took me some time to understand the core functionality. I suppose the main reason for this is the varying terminology of data sources used in vignettes and help pages. The way I understand the logic looks like this:

![medrxiv](https://user-images.githubusercontent.com/309953/85943489-1b44f580-b939-11ea-8dc1-f551da5e0846.png)

In short, for a search target there are two options, the dataset I download myself from medRxiv, or the dataset provided by the GitHub repo. The former can be either all items or just a subset limited by date. The latter is all items. Technically speaking, my download uses the medRxiv API, but the dataset in the repo is built by scraping the medRxiv web site on a daily basis. My understanding is that the main reasons for the scraped dataset are to provide a reliable data source for those occasions when the API does not serve well or not at all, and lighten the burden of the API usage.

How long does it take to download all metadata from the API? I tested it from two physical locations with a differing bandwidth:

```
start_time <- Sys.time()
medrxiv_data <- mx_api_content()
end_time <- Sys.time()
end_time - start_time
```

```
Time difference of 1.434701 mins (1Gb/s line, work)
Time difference of 3.045589 mins (28Mb/s line, home)
```

So far this is not bad, especially if you run the function once a day. One minor thing: is there any way to gracefully stop the process if started by accident? When the RStudio's red Stop button is hit, the following error is thrown

```
Error in curl::curl_fetch_memory(url, handle = handle): Operation was aborted by an application callback
Request failed [ERROR]. Retrying in 1 seconds...
```

`httr::RETRY` is a new function to me. Thanks for this, I will definitely try to use it myself at some point. I wonder though if it allows a clean, user-friendly, forced exit and if yes, how should it be defined?

How rapidly can we expect medRxiv to grow? Looking back, the amount of submissions accelerated when the still very much prevailing COVID-19 pandemic began.

```
library(tidyverse)
library(medrxivr)

mx_data <- mx_api_content()

stats <- mx_data %>%
  mutate(date = as.Date(date)) %>%
  group_by(date) %>%
  summarise(count = n())

png(filename="medrxivstats.png",
    units="cm",
    width=20,
    height=20,
    pointsize=12,
    res=72)

qplot(x=date, y=count,
      data=stats, na.rm=TRUE,
      main="medRxiv item growth",
      xlab="Date of submission", ylab="Number of submissions")

dev.off()
```

![medrxivstats](https://user-images.githubusercontent.com/309953/85943628-ebe2b880-b939-11ea-953d-f7c2fabfc962.png)

Search is a key component of this package, and vignettes help in building search queries. The *medrxivr* one shows how to use the `mx_search` function: either within a two-step process, or with a one-step or piped process. The examples are a little confusing though because the functions shown are not the same; the first example uses `mx_api_content`, the second one `mx_api` which does not exists. I suppose `mx_api` is a typo, maybe the name of a former version? The vignette *building-complex-search-strategies* shows several strategies to filter data, and also how to use regular expressions. Very helpful. One minor thing about this example

```
mx_results <- mx_search(query = "dementia",
                        NOT = "mild cognitive impairment")
```
The NOT argument does not match to _Mild cognitive impairment_ which is found in one abstract, so perhaps better to use the form of `[mM]ild cognitive impairment` instead.

In `mx_search` , the `data` argument is important because it defines the target. Again, the example in the help file is slightly misleading because there is no `mx_raw` function. A former version this one too I presume?

When I ran `mx_search` with zero arguments, my first thought was that there are some issues with error handling. The query starts but clearly you need to include the search string too! However, after some time the error handling kicks in and correctly reminds me of the missing `query` argument. If I am not mistaken, the delay was caused by the latency of the default data source in the GitHub repository.

As of writing this, how long does it take to query the repo?

```
start_time <- Sys.time()
mx_results <- mx_search(query = "molecular")
end_time <- Sys.time()
(end_time - start_time)
```

```
Using medRxiv snapshot - 2020-06-27 06:01
Found 226 record(s) matching your search.
Time difference of 20.75107 secs
```

To me this is acceptable, but people of today tend to be impatient. Still, when the same search against my local copy of the medRxiv database takes only 0.5 secs, you begin to wonder which one to use. I noticed that the question of how to efficiently host and serve a dataset is something you and the editor have already discussed about. Unfortunately, I cannot give any advice, but am very much interested to learn about this topic too. I hope you will find a good solution.

Downloading PDFs works smoothly and as promised. Note: the `mx_download` help file example of `mx_search` uses a `limit` argument which is not defined.

The [Shiny application](https://mcguinlu.shinyapps.io/medrxivr/) that comes with the package is a beautiful piece of work, and the idea of delivering reproducible code is a nice one indeed. However, there are some issues with the code. Both the basic and advanced search codes throw an error when run in R.

Basic:

```
query <- "coronavirus"
mx_results <- mx_search(query)
```
`Error: $ operator is invalid for atomic vectors`

Advanced:

```
topic1 <- c("coronavirus")
topic2 <- c("airborne")
query <- list(topic1, topic2)
mx_results <- mx_search(query, from.date =20190101, to.date =20200628, NOT = c(""), deduplicate = TRUE)
```
`Error in UseMethod("filter_") : no applicable method for 'filter_' applied to an object of class "list"`

I was noted by @maurolepore that the package includes also a short manuscript to be submitted to Journal of Open Source Software. I found the manuscript in the `inst` directory, read it, and found it to be both clear and concise. Good luck!
