### 2020-06-29 #380 @maurolepore
@tts, thanks for your wonderful review!

@mcguinlu, RE

> "Is it better for me to wait until the second reviewer has filed their review before beginning to make changes?"

Both reviewers should work on the exact same package. You may change the package in a separate branch, but please only merge it after both reviewers submitted their review.
