+++ pkgdown.yml
@@ -1,9 +1,8 @@
-pandoc: 2.7.3
-pkgdown: 1.5.1.9000
-pkgdown_sha: ac78596154e403df5f4e683f2185d88225a0fea6
+pandoc: '2.2'
+pkgdown: 1.5.1
+pkgdown_sha: ~
 articles:
-  building-complex-search-strategies: building-complex-search-strategies.html
-  medrxiv-api: medrxiv-api.html
+  intro-to-medrxivr: intro-to-medrxivr.html
   medrxivr: medrxivr.html
-last_built: 2020-05-27T14:20Z
+last_built: 2020-06-29T13:36Z
 
