### 2020-06-29 #380 @tts
> One specific thing I wanted to follow-up on was that the "Automated testing" item in the reviewer checklist is not marked as complete - did you have any specific issues with/reccomendations for this area of the pacakge?

Sorry, my bad. Both `devtools::check()` and `devtools::test()` ran without errors. Checked that item in the list.
