### 2020-07-08 #380 @njahn82
>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is very timely package that not just reflect the increasing popularity of open access preprints in Health Sciences, but also issues around finding and searching them. Although a growing suite of scholarly search engines make medRxiv preprints available, there seems to be no standard way to retrieve  data from medRxiv thoroughly and systematically. Also finding full-texts is challenging, because medRxiv preprints are not made available via PubMed Central. Similiary, Crossref metadata, medRxiv's DOI registration agency, lack links to pdf full-texts.

Before I share my code review, I want to disclose that I neither have an academic background in Health Sciences nor have I been involved in systematic reviews as a librarian. I will therefore focus on more formal aspects of the package and its design.

#### Overall Design

The package contains functions to retrieve metadata from medRxiv, applying complex search strategies on a metadata snapshot, and download pdf full-texts. However, the source code repository contains a considerable amount of other functionality as well, which is outside of the `R` directory and excluded from the package build in `.Rbuildignore`:

- `app` comprises a nice-looking and useful Shiny app helping users to build queries and visually explore the results.
- `data-extraction` has scripts and functions for fetching and validating data.

There's also a link to (daily updated) data in an external GitHub repo, https://github.com/mcguinlu/medrxivr-data/, which is used in an exported R function.

My main concern with this approach is that dependencies, which are not part of the package, are loaded, and [in one case installed](https://github.com/mcguinlu/medrxivr/blob/a299b6965ff676b5375dc34d8948a6d0ef823399/app/app.R#L275). The code outside of the `R` folder also lacks documentation using roxygen tags and tests, and there's some redundancy. I feel that R code not part of the {medrxivr} package build either needs to be factored out should be moved into the `R/` directory.

In the following, I will focus on the functionality, which is part of the package build.

- The README is very helpful to get started with the package. A brief description of what medRxiv is and a link to the preprint server would make the README more informative.
- Maybe  the distinction between downloading a snapshot and searching the remote snapshot could be made a bit clearer. I first started to download the whole corpus, and then realised that there's already a snapshot that I can use instead.
- I love @tts sketch of the overall design. Maybe it can be adapted and re-used?

- In the beginning, I followed the docs on https://mcguinlu.github.io/medrxivr/index.html and faced several issues. It took me a while before realising that the pkgdown website is outdated, because it was not build after code and documentation updates.
- Documentation of functions could be expanded, particularly roxygen tags `@import` and `@importFrom` do not cover all external functions used.
- Documentation of `mx_search()` refers to a function called `mx_raw()`, which is not part of the package.
- Preprint are a quite new scholarly communication phenomena in Health Science and not all health scientists publish preprints regularly. Moreover, some other preprint servers target health science publications. Therefore, I think it would be good to warn users that medRxiv only contains parts of the Health Science (preprint) literature.

#### Vignette

- There are three vignettes, which is great. Again, the general overview misses a sentence about what the preprint server medRxiv is about.
- Not all code chunks are rendered. Some are introduced with [a blank between the ticks and `{r}`](https://github.com/mcguinlu/medrxivr/blob/e281f4b813c2d3aece111e3313866b262768c70a/vignettes/building-complex-search-strategies.Rmd#L37) Is this intentional?

- There is a considerable duplication of code regarding the API call, which can make it hard to update the package in case of API changes. It would be good to have a single function for the API call.
- URL paths are constructed using `paste`. `httr::modify_url()` and the `query` of `httr::GET()` allow passing arguments to a API. Furthermore, {httr} provides helpful functionality to capture API errors more systematically than in the current implementation.
- `mx_crosscheck()` does web scraping, which is fine according to the robots.txt. However, the requested crawl delay of 7 sec has been not implemented, yet.

Here's the checking using [{polite}](https://github.com/dmi3kno/polite)
``` r
polite::bow("https://www.medrxiv.org/archive", force = TRUE)
#> <polite session> https://www.medrxiv.org/archive
#>     User-agent: polite R package - https://github.com/dmi3kno/polite
#>     robots.txt: 68 rules are defined for 1 bots
#>    Crawl delay: 7 sec
#>   The path is scrapable for this user-agent
```

- `mx_search()` returns a grouped tibble. Personally, I prefer to have an ungrouped tibble. The column `date` is of type double, not date.
- Because of the downloading time, it is good to have feedback about the progress. Maybe re-using a progress bar functionality like from [{progress}](https://github.com/r-lib/progress) can lead to less code, while expanding the current feedback mechanism.
- `mx_search()`: rOpenSci style guide recommends snake case for params (`from.date` and `to.date`)
- Finally, I wonder, if Europe PMC could be of use for searching medRxiv. Europe PMC search syntax is quite extensive and supports Boolean operator, wildcards and controlled vocabularies. What are the reasons not using it for searching medRxiv? Is it an indexing lag, or lacking metadata?

Here's a reprex using the vignette example, which took less than 2 second.

``` r
library(tidyverse)
library(europepmc)
ep_q <-
  c('PUBLISHER:"medRxiv" AND (mendelian* AND (randomisation OR randomization))')
epmc_l <- europepmc::epmc_search(ep_q, "raw", limit = 10000)
#> 91 records found, returning 91

my_df <-
  purrr::map_dfr(epmc_l, `[`, c("doi", "title", "abstractText"))
my_df %>%
  filter_at(vars(abstractText, title), any_vars(
    grepl(
      "[Mm]endelian(\\s)([[:graph:]]+\\s){0,4}randomi([[:alpha:]])ation",
      .
    )))
#> # A tibble: 81 x 3
#>    doi           title                         abstractText
#>    <chr>         <chr>                         <chr>
#>  1 10.1101/2020… Cardiometabolic traits, seps… Objectives: To investigate wheth…
#>  2 10.1101/2020… The relationship between gly… Aims: To investigate the relatio…
#>  3 10.1101/2020… Modifiable lifestyle factors… Aims: Assessing whether modifiab…
#>  4 10.1101/2020… Influence of blood pressure … Objectives: To determine whether…
#>  5 10.1101/2020… Increased adiposity is prote… Background Breast and prostate c…
#>  6 10.1101/2020… Examining the association be… Background: We examined associat…
#>  7 10.1101/2020… Investigating the potential … Aim: Use Mendelian randomisation…
#>  8 10.1101/2020… Unhealthy Behaviours and Par… Objective: Tobacco smoking, alco…
#>  9 10.1101/2020… Exploring the causal effect … BACKGROUND: Hearing loss has bee…
#> 10 10.1101/2020… Genetically informed precisi… Impaired lung function is associ…
#> # … with 71 more rows
```

(Disclaimer: I maintain the {europepmc} package and I am curios to learn more about potential shortcomings using Europe PMC instead of a primary literature source. Because I also find it sometimes not very helpful when reviewers point to their own work, I do not expect you to consider this :-))

- All tests passed, but it took a while. My duration was 1221.5 sec. However, I was connected to the internet via a cell phone connection during the review of the package.
- I realised that a lot of skipping for CI platforms happens and I wonder why? Is it the run-time?

I think that's it from me! Thank you for making Health Science preprints more accessible and better discoverable! Happy to help further with the process!
