### 2020-07-09 #380 @mcguinlu
@njahn82 Thanks a million for your detailed review! At a quick skim, everything you flag/recommend is fixable/implementable, and will definitely help to improve the functionality. I'm also looking forward to examing `europepmc` further - to be honest, I was not aware that medRxiv preprints were indexed in Europe PMC.

@maurolepore Just confirming that I have seen this, and so am aiming to address the comments by 23rd July (at the latest).
