### 2020-07-09 70523a5
Move app to own repo
* Remove app folder + content. Now at https://github.com/mcguinlu/medrxivr-app

* Remove app/ from .Rbuildignore
