+++ app/app.R
@@ -1,728 +1,0 @@
-###########################################################################
-# About -------------------------------------------------------------------
-###########################################################################
-#
-# Title: medrxivr web application
-# Author: Luke A McGuinness (luke.mcguinness@bristol.ac.uk)
-# Description: This app was created to be a companion to the medrixvr R package
-# (https://github.com/mcguinlu/medrxivr), which allows systematic reviewers to
-# to use complex search queries to find relevant preprints in the medRxiv
-# (https://www.medrxiv.org/) repository. The app also allows users to explore
-# some simple visualisations of their search results and export their results in
-# a useable format.
-#
-# Other notes: Submitted to the Shiny Contest 2020
-
-###########################################################################
-# External scripts --------------------------------------------------------
-###########################################################################
-
-# Library calls
-source("R/library.R")
-
-# Define function for main trend plot
-source("R/plot_results.R")
-
-# Define function for wordcloud plot
-source("R/plot_word_cloud.R")
-
-# Define misc variables and JS code
-source("R/misc.R")
-
-###########################################################################
-# User interface ----------------------------------------------------------
-###########################################################################
-
-# Define UI
-ui <- tagList(
-  tags$head(includeHTML("www/google-analytics.html")),
-  navbarPage(
-
-        title = "medrxivr",
-        id = "mytabsetpanel",
-        theme = shinythemes::shinytheme("yeti"),
-
-    # Tab with the welcome screen and basic search interface ----
-    tabPanel(
-      "Welcome",
-      includeMarkdown("text/intro.md"),
-      p("To get started, enter a search term (",
-        actionLink("loadbasicsearch","click here to load an example"),
-        ") into the box below or go to the \"Advanced search\" tab to design more complex searches. Your search results are presented as a table and are available for easy download as a comma seperated file (CSV). Summary graphics describing your search results are presented. Finally, as this app is designed as a companion to the ",
-        code("medrxivr"),
-        " R package, the code used to perform your search in R is provided, meaning it is readily reproducible.",        "To see more about the apps background and development, please click ",
-        actionLink("gotomore", "here.")
-      ),
-
-      hr(),
-      h1("Basic search", align = "center"),
-      fluidRow(
-        column(
-          6,
-          align = "center",
-          offset = 3,
-          p(snapshot_info),
-
-          # Simulate button click with "Enter" key
-          # Allows users to hit enter to start basic search
-          tagAppendAttributes(
-            textInput("basicsearchquery", NULL, value = ""),
-            `data-proxy-click` = "basicsearchbutton"
-          ),
-          actionButton("basicsearchbutton", label = "Start basic search"),
-          br(),
-          actionLink("gotoadv", "Advanced search")
-
-        )
-      ),
-
-      # Define initalising loading screen
-      waiter_show_on_load(
-        html = tagList(spin_folding_cube(),
-                       br(),
-                       h4("Initalising medrxivr")),
-        color = "#333333"
-      )
-    ),
-
-    # Tab with the advanced search interface ----
-    tabPanel(
-    "Advanced search",
-    value = "advsearch",
-
-    # Read text from file
-    includeMarkdown("text/advsearch.md"),
-    p(
-      "Click ",
-      actionLink("loadsearch", "here"),
-      " to load an example search (tests for coronavirus) and",
-      " see the ",
[…]
