+++ app/R/library.R
@@ -1,19 +1,0 @@
-# Library calls
-
-library(medrxivr)        # the package this app provides an interface to
-
-library(shiny)           # for app development
-library(waiter)          # for nice loading screens
-library(ggplot2)         # for nice plots
-library(plotly)          # for making plots interactive
-library(htmlwidgets)     # for improving user experience
-library(dplyr)           # for data wrangling
-library(tidytext)        # for natural language processing
-library(stringr)         # to deal with strings (so many strings!)
-library(wordcloud2)      # to render wordclouds
-library(knitr)           # for tables
-library(tidyr)           # for creating tidy data from text
-library(shinyjs)         # to include JavaScript operations in Shiny apps
-library(RColorBrewer)    # to get nice colour palettes
-library(bib2df)          # to export results as .bib file
-library(handlr)
