+++ app/text/intro.md
@@ -1,9 +1,0 @@
-#### Welcome to the medrxivr app
-
-Preprints are full drafts of research papers that are shared publicly before they have undergone [peer-review](https://en.wikipedia.org/wiki/Peer_review), the quality control process where other academics are invited to critically assess the paper and point out flaws/mistakes that may affect its results. The [medRxiv preprint repository](https://medrxiv.org) hosts preprints related to health, and represents a key source of information for those performing a [systematic review](https://en.wikipedia.org/wiki/Systematic_review) - a type of study that aims to find, assess and combine all available knowledge to answer a pressing research question - on a health related topic. This app is designed to allow systematic reviewers to use complex queries, including regular expression syntax and Boolean operators, to search a daily snapshot of the medRxiv repository and download information on preprints related to their research question, something that is currently not possible via the official medRxiv website.
-
-__It is important to remember that preprints should not be relied on to guide clinical practice or health-related behavior and should not be reported in news media as established information.__
-
-<br>
-
-#### Using the app
