+++ app/text/advsearch.md
@@ -1,6 +1,0 @@
-#### Using the advanced search
-
-Enter search terms related to up to 4 topic areas in the boxes below, with each term on its own line. Terms within each topic box are combined with the Boolean operator `OR`, meaning that a record will be retrieved if it includes any of the terms in that box. Different topics are combined using the operator `AND`, meaning for example, that a record will be returned only if it contains at least one of the terms in the  Topic 1 box __and__ at least one of the terms in the Topic 2 box. Enter search terms to exclude in the `NOT` box on the right -  records containing any of these terms will be excluded from those returned. 
-
-Use the options on the left to limit the returned results by date of publication (by default, no restrictions are in place). Finally, it is worth noting that preprints hosted on medRxiv can be updated whenever the authors wish, but a permanet record of all versions is maintainted. Based on this, if you wish to retrieve all versions of a records rather than just to most current one, for example, to explore how the records has changed over time, uncheck the "Remove older versions of the same record?" checkbox.
-

[…]
