### 2020-07-23 e15a25e
add reviewers and new packages to Desc
