### 2020-07-23 3a1490d
Make changes to address rOpenSci review
