+++ vignettes/medrxivr.Rmd
@@ -43,14 +43,14 @@ Alternatively, this can be done all in one step using a piped workflow:
 ``` {r, eval=FALSE}
 library(dplyr)
 
-results <- mx_api() %>%
+results <- mx_api_content() %>%
            mx_search(query ="dementia")
 
 ```
 
 ## Perform a simple search using the daily snapshot of the database
 
-An alternative to the approach detailed above is to search a daily snapshot of the database, taken each morning using the `mx_api_content()` function. Searches performed using this method do not rely on the API (which can become unavilable during peak times) and are usually faster (as it reads data from a CSV rather than having to re-extract it from the API).
+An alternative to the approach detailed above is to search a daily snapshot of the database, taken each morning using the `mx_api_content()` function. Searches performed using this method do not rely on the API (which can become unavailable during peak times) and are usually faster (as it reads data from a CSV rather than having to re-extract it from the API).
 
 ``` {r, eval=FALSE}
 library(medrxivr)
