+++ NAMESPACE
@@ -4,6 +4,7 @@ export(mx_api_content)
 export(mx_api_doi)
 export(mx_crosscheck)
 export(mx_download)
+export(mx_export)
 export(mx_search)
 importFrom(dplyr,"%>%")
 importFrom(methods,is)
