### 2020-07-27 #380 @mcguinlu
Hi all (esp @maurolepore)

A brief message to let you know that I have most of the changes requested made, but due to external circumstances, I haven't yet finished off the small number of outstanding items. I'm now aiming to have it ready for re-review by Thursday week (6th August) at the very latest.

Very sorry for the delay, and hope this is okay!
