### 2020-08-04 390e5e3
Range of changes for rOpenSci review (terrible commit name I know. . .)
