+++ R/mx_download.R
@@ -4,49 +4,69 @@
 #' @param directory The location you want to download the PDF's to
 #' @param create TRUE or FALSE. If TRUE, creates the directory if it doesn't
 #'   exist
+#' @param name How to name the donwloaded PDF. By default, both the ID number of the record and the DOI are used.
 #' @param print_update How frequently to print an update
-#' @examples \dontrun{
-#' mx_results <- mx_search("ecology",limit=20)
-#' mx_download(mx_results,"~/medrxivPDF")
+#' @examples
+#' \dontrun{
+#' mx_results <- mx_search(mx_snapshot(), query = "molecular")
+#' mx_download(mx_results, "medrxiv-PDF")
 #' }
 #' @family main
 #' @export
 #' @importFrom utils download.file
 #' @importFrom methods is
 #' @importFrom stats runif
+#' @importFrom dplyr %>%
 
 
 mx_download <- function(mx_results,
                         directory,
                         create = TRUE,
-                        print_update = 10){
+                        name = c("ID", "DOI"),
+                        print_update = 10) {
+  if (all(c("ID", "DOI") %in% name)) {
+    mx_results$filename <- paste0(mx_results$ID, "_", mx_results$doi)
+  } else {
+    if (name == "ID") {
+      mx_results$filename <- mx_results$ID
+    }
 
-  mx_results$filename <- paste0(mx_results$doi,"v",mx_results$version)
-  mx_results$filename <- gsub("/","-",mx_results$filename)
+    if (name == "DOI") {
+      mx_results$filename <- mx_results$DOI
+    }
+  }
 
-  print(paste0("Estimated time to completion: ",
-               round(length(mx_results$link_pdf)*13/60/60, 2), " hours"))
+  mx_results$filename <- gsub("/", "_", mx_results$filename)
 
-  if(!file.exists(directory)  && create){
+
+  message(paste0(
+    "Estimated time to completion: ",
+    round(length(mx_results$link_pdf) * 13 / 60 / 60, 2), " hours"
+  ))
+
+  if (!file.exists(directory) && create) {
     dir.create(file.path(directory))
   }
 
   # Add trailing forward slash to the directory path
-  if(substr(directory,nchar(directory), nchar(directory)) != "/"){
-    directory <- paste(directory,"/",sep="")
+  if (substr(directory, nchar(directory), nchar(directory)) != "/") {
+    directory <- paste(directory, "/", sep = "")
   }
 
   number <- 1
 
   for (file_location in mx_results$link_pdf) {
-    if (file.exists(paste0(directory,
-                           mx_results$filename[which(mx_results$link_pdf ==
-                                                     file_location)],
-                           ".pdf"))) {
-      message(paste0("PDF for ID ",
-                     mx_results$filename[which(mx_results$link_pdf ==
-                                                                file_location)],
-                     " already downloaded."))
+    if (file.exists(paste0(
+      directory,
+      mx_results$filename[which(mx_results$link_pdf ==
+        file_location)],
+      ".pdf"
+    ))) {
+      message(paste0(
+        "PDF already downloaded for DOI: ",
+        mx_results$filename[which(mx_results$link_pdf ==
+          file_location)]
+      ))
 
       number <- number + 1
 
@@ -54,14 +74,16 @@ mx_download <- function(mx_results,
     }
 
     while (TRUE) {
-      message(paste0("Downloading PDF ",
-                   number,
-                   " of ",
-                   length(mx_results$link_pdf),
-                   " (DOI: ",
-                   mx_results$filename[which(mx_results$link_pdf ==
-                                                       file_location)],
-                   "). . . "))
[…]
