+++ man/api_to_df.Rd
@@ -17,7 +17,7 @@ Concert API data to data frame
 }
 \examples{
 \dontrun{
-df <- api_to_df(api_link("medrxiv","10.1101/2020.02.25.20021568"))
+df <- api_to_df(api_link("medrxiv", "10.1101/2020.02.25.20021568"))
 }
 }
 \keyword{Internal}
