+++ man/mx_search.Rd
@@ -45,10 +45,11 @@ Search medRxiv using a string
 \dontrun{
 # Using data from the API
 mx_results <- mx_api_content() \%>\%
-mx_search(query = "dementia")
+  mx_search(query = "dementia")
 
 # Using the daily snapshot
-mx_results <- mx_search(query = "dementia")
+mx_results <- mx_snapshot() \%>\%
+  mx_search(query = "dementia")
 }
 }
 \seealso{
