+++ man/mx_crosscheck.Rd
@@ -7,7 +7,9 @@
 mx_crosscheck()
 }
 \description{
-Cross-check whether the dataset you are using is current
+Provides information on how up-to-date the snapshot is by
+  checking whether there have been any records added to/updated in the
+  medRxiv repository since the last snapshot was taken.
 }
 \examples{
 \dontrun{
