+++ man/mx_download.Rd
@@ -35,6 +35,7 @@ mx_download(mx_results, "medrxiv-PDF")
 }
 \seealso{
 Other main: 
+\code{\link{mx_export}()},
 \code{\link{mx_search}()}
 }
 \concept{main}

[…]
