### 2020-08-06 e14b53e
Fix vignette issue
