### 2020-08-06 ce7e605
Range of fixes and final changes for review
