+++ R/mx_api.R
@@ -1,4 +1,4 @@
-#' Access medRxiv API
+#' Access Cold Springs Harbour API, which serves both medRxiv and bioRxiv
 #'
 #' @param from_date Earliest date of interest. Defaults to 1st June 2019
 #'   (earliest medRxiv record was posted on 25th June 2019).
@@ -9,7 +9,7 @@
 #' @param server Specify the server you wish to use: "medrxiv" (default) or
 #'   "biorxiv"
 #' @param clean Logical, indicating whether to clean the data returned for use
-#'   with other mx_* functions.
+#'   with other mx_* functions. Default is TRUE.
 #'
 #' @return Dataframe with 1 record per row
 #'
