+++ vignettes/medrxivr.Rmd
@@ -35,13 +35,13 @@ knitr::opts_chunk$set(
 
   - `mx_api_content()` creates a local copy of all data available from the medRxiv API at the time the function is run.
   
-  - `mx_snapshot()` provides acces to a static snapshot of the database. The snapshot is created each morning at 6am using `mx_api_content()` and is stored as CSV file in the [medrxivr-data repository](https://github.com/mcguinlu/medrxivr-data). This method does not rely on the API (which can become unavailable during peak usage times) and is usually faster (as it reads data from a CSV rather than having to re-extract it from the API). Discrepancies between the most recent static snapshot and the live database can be assessed using `mx_crosscheck()`.
+  - `mx_snapshot()` provides access to a static snapshot of the database. The snapshot is created each morning at 6am using `mx_api_content()` and is stored as CSV file in the [medrxivr-data repository](https://github.com/mcguinlu/medrxivr-data). This method does not rely on the API (which can become unavailable during peak usage times) and is usually faster (as it reads data from a CSV rather than having to re-extract it from the API). Discrepancies between the most recent static snapshot and the live database can be assessed using `mx_crosscheck()`.
   
 The relationship between the two methods is summarised in the figure below:
   
 ``` {r eval = TRUE, echo = FALSE, out.width = "500px", out.height = "400px"}
 
-knitr::include_graphics("../figs/data_sources.png")
+knitr::include_graphics("data_sources.png")
 ```
   
 ## Perform a simple search
