+++ .github/workflows/pkgdown.yml
@@ -36,10 +36,12 @@ jobs:
 
       - name: Install package
         run: R CMD INSTALL .
-
-      - name: Deploy package
+        
+      - name: Git Credentials
         run: |
           git config --local user.email "actions@github.com"
           git config --local user.name "GitHub Actions"
-          pkgdown::deploy_to_branch(new_process = FALSE)
+
+      - name: Deploy package
+        run: pkgdown::deploy_to_branch(new_process = FALSE)
         shell: Rscript {0}
