+++ pkgdown.yml
@@ -1,8 +1,6 @@
-pandoc: '2.2'
-pkgdown: 1.5.1
-pkgdown_sha: ~
-articles:
-  intro-to-medrxivr: intro-to-medrxivr.html
-  medrxivr: medrxivr.html
-last_built: 2020-06-29T13:36Z
+pandoc: 2.7.3
+pkgdown: 1.5.1.9000
+pkgdown_sha: eae56f08694abebf93cdfc0dd8e9ede06d8c815f
+articles: []
+last_built: 2020-08-06T16:22Z
 
