### 2020-08-08 #380 @maurolepore
@mcguinlu, thanks for your hard work. See my comments below.

@tts and @njahn82 please consider @mcguinlu 's changes and respond with either your approval or a suggestion for improvement.

--

### Following up previous comments (editor checks)

* ml3: `spelling::spell_check_package()` still shows some unknown words. Please update your words list and
consider automating the process with `usethis::use_spell_check()`.

``` r
spelling::spell_check_package()
#>   WORD      FOUND IN
#> bioRxiv   mx_api_content.Rd:5,35
#> Harbour   mx_api_content.Rd:5,35
```

* ml4: `goodpractice::gp()` still suggests some improvements:

``` r
suppressWarnings(goodpractice::gp())
...
─────────────────────────────────────────────────────────────────
#>
#> It is good practice to
#>
#>   ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R
#>     users and developers are used it and it is easier to read your code
#>     for them if you use '<-'.
#>
#>     tests/testthat/test-helpers.R:4:7
#>     tests/testthat/test-helpers.R:20:7
#>
#>   ✖ avoid long code lines, it is bad for readability. Also, many people
#>     prefer editor windows that are about 80 characters wide. Try make
#>     your lines shorter than 80 characters
#>
#>     R/mx_crosscheck.R:24:1
#>     R/mx_download.R:7:1
#>     R/mx_search.R:37:1
#>     tests/testthat/test-export.R:9:1
#>
#> ────────────────────────────────────────────────────────────────────────────────
```

* ml5: `covr::package_coverage()` shows greater coverage than before; thanks. The only file that's still
a little low is R/mx_crosscheck.R. Please consider adding more tests or excluding code as
necessary (https://github.com/r-lib/covr#exclusions).

```r
# > covr::package_coverage()
# medrxivr Coverage: 91.04%
# R/mx_crosscheck.R: 69.23%
# R/helpers.R: 83.33%
# R/mx_search.R: 90.27%
# R/mx_download.R: 93.85%
# R/mx_api.R: 100.00%
# R/mx_export.R: 100.00%
# R/mx_info.R: 100.00%
# R/mx_snapshot.R: 100.00%
```

* ml6: `usethis::use_tidy_style()` suggests some files could improve. Please run
`usethis::use_tidy_style()` and consider committing the changes.

``` r
usethis::use_tidy_style()
#> ✔ Setting active project to '/home/mauro/git/medrxivr'
#> Styling  17  files:
#>  R/helpers.R                      ℹ
#>  R/medrxivr.R                     ✔
#>  R/mx_api.R                       ✔
#>  R/mx_crosscheck.R                ✔
#>  R/mx_download.R                  ✔
#>  R/mx_export.R                    ✔
#>  R/mx_info.R                      ✔
#>  R/mx_search.R                    ✔
#>  R/mx_snapshot.R                  ✔
#>  tests/testthat.R                 ✔
#>  tests/testthat/test-api.R        ✔
#>  tests/testthat/test-crosscheck.R ✔
#>  tests/testthat/test-download.R   ✔
#>  tests/testthat/test-export.R     ✔
#>  tests/testthat/test-helpers.R    ℹ
#>  tests/testthat/test-info.R       ✔
#>  tests/testthat/test-search.R     ✔
#> ────────────────────────────────────────
#> Status   Count   Legend
#> ✔    15  File unchanged.
#> ℹ    2   File changed.
#> ✖    0   Styling threw an error.
#> ────────────────────────────────────────
#> Please review the changes carefully!
#>
#> ✔ Styled project according to the tidyverse style guide
```

### New comments (suggestions)

* ml7: On the website, the Reference tab shows "All functions". Maybe you can
help users navigate this reference by grouping functions in some meaningful
way? (see https://pkgdown.r-lib.org/reference/build_reference.html).

* ml8: You may want to consider setting up a CI services for a wider range of
environments. Here are two workflows you may use:

    * https://usethis.r-lib.org/reference/use_github_action.html#use-github-action-check-standard-
    * https://usethis.r-lib.org/reference/use_github_action.html#use-github-action-check-full-

* ml9: I see three .Rmd files inside vignettes/ but only two in the Articles section
of the website. Is this itentional?  Also, vignettes are great, but they can
make the installation heavier.  Consider the difference between use_vignette()
and use_article(): https://usethis.r-lib.org/reference/use_vignette.html

* ml10: I recommend walking through the steps listed by
[`use_release_issue()`](https://usethis.r-lib.org/reference/use_release_issue.html)
or [`devtools::release()`](https://devtools.r-lib.org/reference/release.html).
Even if you don't submit to CRAN, walking through the process can help you find
details to improve.

* ml11: The vignettes show code but not output. Reproducible examples are most useful
when they include the output because readers can understand what the code does
even if they choose not to run the code themselves. This is why `reprex::reprex()` prints
output (https://reprex.tidyverse.org/).
