### 2020-08-10 #380 @tts
Hi @mcguinlu and thanks for your efforts! Below, I'll use your headings, and give my remarks to each of them.

> General comments.
In addition to adding your suggested diagram, I have tried to make the language used across the documentation more consistent, but please do point out anything that could be clearer!

Great, much better now. I cannot find anything more to complain :)

> API.
having to hit "Esc"/click "Stop" multiple times in order to actually get mx_api_content() to stop. Is this right?

Yes, that's what I mean, and I understand your explanation. Several clicks do terminate the download, so I find this sufficient.

> Snapshot.
Trying mx_search(query="molecular")

`Time difference of 6.319079 secs` for me, which is not bad.

> Vignette/README

Ok now.

> Download

Ok.

> Shiny app
mx_search()

Reproducible code works now fine, and a missing `data | query` argument is caught right away. Good!

**One new comment**

`mx_info(commit = "master")`
`Error in mx_info(commit = "master") : could not find function "mx_info"`

Except this comment, I give my approval.
