### 2020-08-10 a336a99
Fix good practice issues
