+++ tests/testthat/test-export.R
@@ -6,7 +6,9 @@ if (file.exists("medrxiv_export.bib") == TRUE) {
   unlink("medrxiv_export.bib", recursive = TRUE)
 }
 
-mx_result <- mx_search(mx_snapshot(), query = c("dementia"), to_date = "2020-01-01")
+mx_result <- mx_search(mx_snapshot(),
+                       query = c("dementia"),
+                       to_date = "2020-01-01")
 
 test_that("Inital output", {
   skip_if_offline()
