+++ tests/testthat/test-helpers.R
@@ -1,7 +1,7 @@
 test_that("API to Dataframe",{
   skip_if_offline()
   skip_if_api_message()
-  url = "https://api.biorxiv.org/details/medrxiv/2020-06-21/2020-08-28/45"
+  url <- "https://api.biorxiv.org/details/medrxiv/2020-06-21/2020-08-28/45"
   df <- api_to_df(url)$collection
   expect_identical(class(df),"data.frame")
 })
@@ -17,6 +17,6 @@ test_that("Link modification",{
 test_that("No posts found - date range",{
   skip_if_offline()
   skip_if_api_message()
-  url = "https://api.biorxiv.org/details/medrxiv/2018-06-21/2018-08-28"
+  url <- "https://api.biorxiv.org/details/medrxiv/2018-06-21/2018-08-28"
   expect_error(api_to_df(url))
 })
