### 2020-08-10 49da786
Update to use tidy style
