### 2020-08-10 59cb2a2
Change "Internal" to "internal" so pkgdown site builds correctly
