+++ R/helpers.R
@@ -1,7 +1,7 @@
 #' Checks whether the user has internet, and returns a helpful message it not.
 #'
 #' @return Informative error if not connected to the internet
-#' @keywords Internal
+#' @keywords internal
 
 internet_check <- function() {
   if (curl::has_internet() == FALSE) {
@@ -17,7 +17,7 @@ internet_check <- function() {
 #' @param url API endpoint from which to extract and format data
 #'
 #' @return Raw API data in a dataframe
-#' @keywords Internal
+#' @keywords internal
 #'
 #' @examples
 #' \dontrun{
@@ -71,7 +71,7 @@ api_to_df <- function(url) {
 #' @param ... Arguments to specify the path to the API endpoint
 #'
 #' @return Formatted link to API endpoint
-#' @keywords Internal
+#' @keywords internal
 #'
 #' @examples
 #' \dontrun{
@@ -95,7 +95,7 @@ api_link <- function(...) {
 #' @param df Raw dataframe from API
 #'
 #' @return Cleaned dataframe
-#' @keywords Internal
+#' @keywords internal
 #'
 #' @examples
 #' \dontrun{
