+++ R/mx_info.R
@@ -3,7 +3,7 @@
 #' @param commit Commit hash for the snapshot, taken from
 #'   https://github.com/mcguinlu/medrxivr-data. Defaults to "master", which will
 #'   return info on the most recent snapshot.
-#' @keywords Internal
+#' @keywords internal
 #' @return Message with snapshot details
 #'
 #' @examples
