+++ R/mx_crosscheck.R
@@ -1,7 +1,7 @@
-#' Live check against medRxiv
-#' @description Provides information on how up-to-date the snapshot is by
-#'   checking whether there have been any records added to/updated in the
-#'   medRxiv repository since the last snapshot was taken.
+#' Cross-check the medRxiv snapshot against the Cold Springs Harbour API
+#' @description Provides information on how up-to-date the maintained medRxiv
+#'   snapshot is by checking whether there have been any records added
+#'   to/updated in the medRxiv repository since the last snapshot was taken.
 #' @examples
 #' \dontrun{
 #' mx_crosscheck()
