### 2020-08-10 f77b7ff
Add better build instructions to "Reference" section of the pkgdown website
