+++ _pkgdown.yml
@@ -1,1 +1,13 @@
 destination: docs
+reference:
+- title: Data sources
+  desc:  Access preprint data
+  contents:
+  - has_concept("data-source")
+- title: Main functions
+  desc:  Work with imported preprint data
+  contents:
+    - has_concept("main")
+- title: Helper functions
+  contents:
+  - has_concept("helper")
