### 2020-08-10 4ec2254
Add CI badge to README
