### 2020-08-10 1dd43c9
Minor changes to vignettes
