### 2020-08-11 f23ed0b
Update vignettes
