### 2020-08-11 39a278c
Clean-up and trigger CI
