+++ man/mx_crosscheck.Rd
@@ -2,14 +2,14 @@
 % Please edit documentation in R/mx_crosscheck.R
 \name{mx_crosscheck}
 \alias{mx_crosscheck}
-\title{Live check against medRxiv}
+\title{Cross-check the medRxiv snapshot against the Cold Spring Harbour Laboratory API}
 \usage{
 mx_crosscheck()
 }
 \description{
-Provides information on how up-to-date the snapshot is by
-  checking whether there have been any records added to/updated in the
-  medRxiv repository since the last snapshot was taken.
+Provides information on how up-to-date the maintained medRxiv
+  snapshot is by checking whether there have been any records added
+  to/updated in the medRxiv repository since the last snapshot was taken.
 }
 \examples{
 \dontrun{
