+++ rOpenSci_response.md
@@ -1,212 +1,55 @@
-# Overview
 
-To start, thanks once again to everyone for your constructive comments! I've gone through all the posts above and have tried to extract all feedback here so that I can address it systematically. - please let me know if I have missed anything. Comments are divided by commenter and topic for ease of reference, and are presented in bold text, with the response immediately below.
-
-I have also acknowledged both @tss and @njahn82 as reviewers in the `Description`, as I feel like your comments have added a great deal to the functionality/user-friendliness of the package. Please check your details just to make sure I have them correct!
-
-<hr>
-
-# Some key points about new functionality:
-
-- [x] I've added a new function (`mx_export()`) to save search results as a .bib file for import to reference management software (this functionality already existed for the app but hadn't been copied across to the package).
-
-- [x] I have seperated out importing the snapshot into a seperate function (`mx_snapshot()`), rather than just using `mx_search(data= NULL)` to indicate that you want to use the snapshot. I think this will help users be very clear about which data source they are using, and was influenced by feedback from both reviewers about being confused re: my description of the data sources. 
-
-- [x] Due to the fact that the API endpoints are identical for both medRxiv and bioRxiv and the fact that all interactions with the API have been centralised into a single function (thanks to reviewers feedback), coupled with demand from others in my group/people emailing me about it, the `mx_api_*()` functions now contain a `server` argument, allowing users to specify which server (medRxiv or bioRxiv) they want to interact with. Downloading the whole bioRxiv database does take a while, but users seem happy to do so rather than having to perform the searches and download the results manually via the website ([see here for an example](https://github.com/abannachbrown/MedrxivBiorxivSearches)). One thing I am worried about is whether to keep this as a happy bonus of the packages restructure following the reviews above, or whether to publicise it further/make it a key element of the package? Any advice/comments from reviewers/editors would be apprecited here!
-
-# Editor (@maurolepore)
-
-## Tasks 
-
-These have all been completed:
-
-- [x] Add rOpenSci badge to README
-- [x] Fix spelling (ml3)
-- [x] Fix style (ml6)
-
-## Discussion points
-
-- [x] **How big is the database? How fast does it grow? And how long does it take to download it in a range of reasonable conditions? What happens in a range of extreme conditions?**  
-As mentioned by @tss, there has been a substantial surge in the size of the database thanks to the COVID-19 pandemic. @tss reported times of 1.4 and 3.1 mins to download the database. I am not sure what is meant by extreme circumstances, but happy to do more testing if needed!
-
-- [x] **Is the process transparent and "polite" to the user?**  
-I'm not 100% sure what was meant by this comment. Re: interaction with medRxiv, all information is now taken from the API (previously `mx_crosscheck()` used web-scraping, but this is no longer the case) - is this what was meant?
-
-<hr>
-
-# Reviewer 1 (@tts)
-
-## General comments
-
-- [x] **Although the target group and the goal of the package are clearly defined, it took me some time to understand the core functionality. I suppose the main reason for this is the varying terminology of data sources used in vignettes and help pages.**  
-In addition to adding your suggested diagram, I have tried to make the language used across the documentation more consistent, but please do point out anything that could be clearer!
+* **ml3: `spelling::spell_check_package()` still shows some unknown     words. Please update your words list and consider automating the process with `usethis::use_spell_check()`.**
+  I have automated the spellcheck now as recommended.
 
 
-## API
[…]
