### 2020-08-14 9768ce1
Major changes to README and Get Started vignette
