+++ README.md
@@ -20,23 +20,25 @@ status](https://ci.appveyor.com/api/projects/status/github/mcguinlu/medrxivr?bra
 coverage](https://codecov.io/gh/mcguinlu/medrxivr/branch/master/graph/badge.svg)](https://codecov.io/gh/mcguinlu/medrxivr?branch=master)
 <!-- badges: end -->
 
-[medRxiv](https://www.medrxiv.org/) is a free online repository for
-complete but unpublished manuscripts (preprints) in the medical,
-clinical, and related health sciences. The `medrxivr` package provides
-programmatic access to metadata (title, authors, date, etc.) on the
-preprints contained in the repository via the [medRxiv
-API](https://api.biorxiv.org/) or a maintained static snapshot (see
-[Data sources](#data-sources).) `medrxivr` also provides functions to
-search medRxiv records using regular expressions and Boolean logic, in
-addition to helper functions to export your search results to a .BIB
-file for easy import to a reference manager and to download the
-full-text PDFs of relevant preprints.
+An increasingly important source of health-related bibliographic content
+are preprints - preliminary versions of research articles that have yet
+to undergo peer review. The two preprint repositories most relevant to
+health-related sciences are [medRxiv](https://www.medrxiv.org/) and
+[bioRxiv](https://www.biorxiv.org/), both of which are operated by the
+Cold Spring Harbor Laboratory.
 
-**Note:** `medrxivr` is now available as a web-app, which lets you build
-complex searches via a user-friendly interface, explore the results and
-export them for screening. In an effort to improve reproducibility, it
-also creates the code needed to run the search straight from R. The app
-is available [here.](https://mcguinlu.shinyapps.io/medrxivr/)
+The goal of the `medrxivr` R package is two-fold. In the first instance,
+it provides programmatic access to the [Cold Spring Harbour Laboratory
+(CSHL) API](https://api.biorxiv.org/), allowing users to easily download
+medRxiv and bioRxiv preprint metadata (e.g. title, abstract, publication
+date, author list, etc) into R. The package also provides access to a
+maintained static snapshot of the medRxiv repository (see [Data
+sources](#medrxiv-data)). Secondly, `medrxivr` provides functions to
+search the downloaded preprint records using regular expressions and
+Boolean logic, as well as helper functions that allow users to export
+their search results to a .BIB file for easy import to a reference
+manager and to download the full-text PDFs of preprints matching their
+search criteria.
 
 ## Installation
 
@@ -47,16 +49,23 @@ devtools::install_github("mcguinlu/medrxivr")
 library(medrxivr)
 ```
 
-## Usage
+## Data sources
 
-### Data sources
+### medRxiv data
 
 `medrixvr` provides two ways to access medRxiv data:
 
-  - `mx_api_content()` creates a local copy of all data available from
-    the medRxiv API at the time the function is run.
+  - `mx_api_content(server = "medrxiv")` creates a local copy of all
+    data available from the medRxiv API at the time the function is run.
 
-  - `mx_snapshot()` provides access to a static snapshot of the
+<!-- end list -->
+
+``` r
+# Get a copy of the database from the live medRxiv API endpoint
+preprint_data <- mx_api_content()  
+```
+
+  - `mx_snapshot()` provides access to a static snapshot of the medRxiv
     database. The snapshot is created each morning at 6am using
     `mx_api_content()` and is stored as CSV file in the [medrxivr-data
     repository](https://github.com/mcguinlu/medrxivr-data). This method
@@ -66,44 +75,65 @@ library(medrxivr)
     between the most recent static snapshot and the live database can be
     assessed using `mx_crosscheck()`.
 
-The relationship between the two methods is summarised in the figure
[…]
