+++ vignettes/medrxivr.Rmd
@@ -25,73 +25,74 @@ knitr::opts_chunk$set(
 
 ```
 
-[medRxiv](https://www.medrxiv.org/) is a free online repository for complete but unpublished manuscripts (preprints) in the medical, clinical, and related health sciences. The `medrxivr` package provides programmatic access to the [medRxiv API](https://api.biorxiv.org/), in addition to a static snapshot of the _medRxiv_ preprint repository which is automatically updated each morning (see [Data sources](#data-sources)). 
+An increasingly important source of health-related bibliographic content are preprints - preliminary versions of research articles that have yet to undergo peer review. The two preprint repositories most relevant to health-related sciences are [medRxiv](https://www.medrxiv.org/) and [bioRxiv](https://www.biorxiv.org/), both of which are operated by the Cold Spring Harbor Laboratory.
 
-## Usage
+The goal of the `medrxivr` R package is two-fold. In the first instance, it provides programmatic access to the [Cold Spring Harbour Laboratory (CSHL) API](https://api.biorxiv.org/), allowing users to easily download medRxiv and bioRxiv preprint metadata (e.g. title, abstract, publication date, author list, etc) into R. The package also provides access to a maintained static snapshot of the medRxiv repository (see [Data sources](#medrxiv-data)). Secondly, `medrxivr` provides functions to search the downloaded preprint records using regular expressions and Boolean logic, as well as helper functions that allow users to export their search results to a .BIB file for easy import to a reference manager and to download the full-text PDFs of preprints matching their search criteria.
 
-### Data sources
+## Installation
+
+You can install the development version of this package using:
+
+``` {r}
+devtools::install_github("mcguinlu/medrxivr")
+library(medrxivr)
+```
+
+## Data sources
+
+### medRxiv data
 
 `medrixvr` provides two ways to access medRxiv data:  
 
-  - `mx_api_content()` creates a local copy of all data available from the medRxiv API at the time the function is run.
+  - `mx_api_content(server = "medrxiv")` creates a local copy of all data available from the medRxiv API at the time the function is run.
   
-  - `mx_snapshot()` provides access to a static snapshot of the database. The snapshot is created each morning at 6am using `mx_api_content()` and is stored as CSV file in the [medrxivr-data repository](https://github.com/mcguinlu/medrxivr-data). This method does not rely on the API (which can become unavailable during peak usage times) and is usually faster (as it reads data from a CSV rather than having to re-extract it from the API). Discrepancies between the most recent static snapshot and the live database can be assessed using `mx_crosscheck()`.
+``` {r}
+# Get a copy of the database from the live medRxiv API endpoint
+preprint_data <- mx_api_content()  
+```
   
-The relationship between the two methods is summarised in the figure below:
+  - `mx_snapshot()` provides access to a static snapshot of the medRxiv database. The snapshot is created each morning at 6am using `mx_api_content()` and is stored as CSV file in the [medrxivr-data repository](https://github.com/mcguinlu/medrxivr-data). This method does not rely on the API (which can become unavailable during peak usage times) and is usually faster (as it reads data from a CSV rather than having to re-extract it from the API). Discrepancies between the most recent static snapshot and the live database can be assessed using `mx_crosscheck()`.
+  
+``` {r}
+# Get a copy of the database from the daily snapshot
+preprint_data <- mx_snapshot()  
+```
+  
+The relationship between the two methods for the medRxiv database is summarised in the figure below:
   
 ``` {r eval = TRUE, echo = FALSE, out.width = "500px", out.height = "400px"}
 
-knitr::include_graphics("data_sources.png")
[…]

[…]
