+++ inst/paper.md
@@ -23,12 +23,11 @@ bibliography: paper.bib
 
 # Summary
 
-An increasingly imporant source of health-related bibliographic content is the medRxiv repository, a free online archive and distribution server for preprints (preliminary versions of research articles that have yet to undergo peer review) in the medical, clinical, and health-related sciences [@rawlinson2019]. Founded in June 2019, the repository has now grown to over 5500 preprints, helped in part by the expotential growth in preprints related to the COVID-19 pandemic. 
+An increasingly important source of health-related bibliographic content are preprints - preliminary versions of research articles that have yet to undergo peer review. The two preprint repositories most relevant to health-related sciences are medRxiv and bioRxiv, both of which are operated by the Cold Spring Harbor Laboratory, a not-for-profit research and educational institution [@rawlinson2019].
 
-The goal of the `medrxivr` R package is two-fold. In the first instance, it provides programmatic access to Cold Spring Harbour Laboratory (CSHL) API, allowing access to both medRxiv and bioRxiv preprint metadata (e.g. title, abstract, authors) from R. Users can either query the [CSHL API](https://api.biorxiv.org/) directly, or can choose to use a daily static snapshot of the database, created and maintained in an effort to limit the package's burden on the API. This functionality will be of interest to anyone who wishes to import medRxiv preprint metadata into R, for example to explore the distribution of preprints by subject area or by publication year. Examples of this type of usage have already been reported [e.g. @Brierley].
+The goal of the `medrxivr` R package is two-fold. In the first instance, it provides programmatic access to the Cold Spring Harbour Laboratory (CSHL) API, allowing users to download medRxiv and bioRxiv preprint metadata (e.g. title, abstract, author list). This functionality will be of interest to anyone who wishes to import medRxiv and/or bioRxiv preprint metadata into R, for example to explore the distribution of preprints by subject area or by publication year. Examples of this type of usage have already been reported [e.g. @Brierley].
 
-In the second instance, the package provides functions that allow users to search the medRxiv database for relevant preprints using complex search strings, incorporating the functionality of search term truncation, Boolean operators (AND, OR, NOT), and term proximity. A helper function is also provided that allows users to easily download the full-text PDFs of preprints matching their search. This aspect of the package will be more relevant to systematic reviewers, health librarians and others performing literature searches, allowing them to perform and document transparent and reproducible searches in this important evidence source.
-
+In the second instance, the package provides functions that allow users to search the downloaded preprint metadata for relevant preprints using complex search strings, including functionality such as search term truncation, Boolean operators (AND, OR, NOT), and term proximity. Helper functions are provided that allow users to export the results of their search to a .bib file for import into a reference manager (e.g. Zotero) and to download the full-text PDFs of preprints matching their search. This aspect of the package will be more relevant to systematic reviewers, health librarians and others performing literature searches, allowing them to perform and document transparent and reproducible searches in these important evidence sources.
 
 # Acknowledgements
 
