+++ tests/testthat/test-api.R
@@ -39,7 +39,12 @@ test_that("Check number of columns in include.info output", {
 })
 
 test_that("Check number of columns in output", {
-  skip_if_offline()
-  skip_if_api_message()
+
   expect_equal(ncol(mx_api_doi("10.1101/2020.02.25.20021568")), 14)
+})
+
+test_that("Server not recognised", {
+  skip_if_offline()
+  expect_error(mx_api_content(server = "medRxiv"))
+  expect_error(mx_api_doi(server = "medRxiv"))
 })
