+++ R/mx_crosscheck.R
@@ -1,7 +1,10 @@
-#' Cross-check the medRxiv snapshot against the Cold Spring Harbour Laboratory API
+#' Check how up-to-date the maintained medRxiv snapshot is
+#'
 #' @description Provides information on how up-to-date the maintained medRxiv
-#'   snapshot is by checking whether there have been any records added
-#'   to/updated in the medRxiv repository since the last snapshot was taken.
+#'   snapshot provided by `mx_snapshot()` is by checking whether there have been
+#'   any records added to, or updated in, the medRxiv repository since the last
+#'   snapshot was taken.
+#'
 #' @examples
 #' \dontrun{
 #' mx_crosscheck()
