+++ R/medrxivr.R
@@ -1,7 +1,10 @@
-#' medrxivr: Accessing medRxiv data from R
+#' medrxivr: Accessing medRxiv and bioRxiv preprint data from R
 #'
-#' The medrxivr package enables users to access data from the medRxiv
-#' preprint repository from R
+#' The medrxivr package enables users to access data on preprints in the medRxiv
+#' and bioRxiv preprints repositories, both of which are run by the Cold Spring
+#' Harbour Laboratory.It also provides functions to search the preprint data,
+#' export it to a .bib file, and download the PDFs asscoiated with specified
+#' records.
 #' @docType package
 #' @name medrxivr
 
