+++ R/mx_info.R
@@ -1,4 +1,4 @@
-#' Provide information on the snapshot used to perform the search
+#' Provide information on the medRxiv snapshot used to perform the search
 #'
 #' @param commit Commit hash for the snapshot, taken from
 #'   https://github.com/mcguinlu/medrxivr-data. Defaults to "master", which will
