### 2020-08-14 dfd64df
use_tidy_style changes
