+++ R/medrxivr.R
@@ -3,7 +3,7 @@
 #' The medrxivr package enables users to access data on preprints in the medRxiv
 #' and bioRxiv preprints repositories, both of which are run by the Cold Spring
 #' Harbour Laboratory.It also provides functions to search the preprint data,
-#' export it to a .bib file, and download the PDFs asscoiated with specified
+#' export it to a .bib file, and download the PDFs associated with specified
 #' records.
 #' @docType package
 #' @name medrxivr
