+++ tests/testthat/test-api.R
@@ -14,7 +14,7 @@ test_that("Check data inputs return the same no. of results", {
     mx_api_content(
       from_date = "2019-01-01",
       to_date = "2020-01-01",
-      include.info = TRUE
+      include_info = TRUE
     )
 
   mx2 <-
@@ -26,14 +26,14 @@ test_that("Check data inputs return the same no. of results", {
   expect_equal(nrow(mx1), nrow(mx2))
 })
 
-test_that("Check number of columns in include.info output", {
+test_that("Check number of columns in include_info output", {
   skip_if_offline()
   skip_if_api_message()
   mx_data <-
     mx_api_content(
       from_date = "2019-01-01",
       to_date = "2020-01-01",
-      include.info = TRUE
+      include_info = TRUE
     )
   expect_equal(ncol(mx_data), 20)
 })
