### 2020-08-14 0489400
Updated the package title in the DESCRIPTION
