+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: medrxivr
-Title: Access MedRxiv Preprint Data
+Title: Access and Search MedRxiv and BioRxiv Preprint Data In R
 Version: 0.0.2.900
 Authors@R: c(
     person("Luke", "McGuinness",
