+++ README.Rmd
@@ -120,13 +120,26 @@ mx_export(data = results,
 
 Pass the results of your search above (the `results` object) to the `mx_download()` function to download a copy of the PDF for each record found by your search.
 
-```{r, eval = FALSE}
+```{r}
 
 mx_download(results,        # Object returned by mx_search(), above
             "pdf/",         # Directory to save PDFs to 
             create = TRUE)  # Create the directory if it doesn't exist
 
 ```
+
+## Accessing the raw API data
+
+By default, the `mx_api_*()` functions clean the data returned by the API for use with other `medrxivr` functions. 
+
+To access the raw data returned by the API, the `clean` argument should  set to `FALSE`:
+
+``` {r}
+mx_api_content(to_date = "2019-07-01", clean = TRUE)
+
+```
+
+See [this article](https://mcguinlu.github.io/medrxivr/articles/medrxiv-api.html#accesing-the-raw-api-data) for more details.
 
 ## Detailed guidance
 
