+++ man/mx_api_doi.Rd
@@ -14,8 +14,12 @@ data on.}
 \item{server}{Specify the server you wish to use: "medrxiv" (default) or
 "biorxiv"}
 
-\item{clean}{Logical, indicating whether to clean the data returned for use
-with other mx_* functions.}
+\item{clean}{Logical, defaulting to TRUE, indicating whether to clean the
+data returned by the API. If TRUE, variables containing absolute paths to
+the preprints web-page ("link_page") and PDF ("link_pdf") are generated
+from the "server", "DOI", and "version" variables returned by the API. The
+"title", "abstract" and "authors" variables are converted to title case.
+Finally, the "type" and "server" variables are dropped.}
 }
 \value{
 Dataframe containing details on the preprint identified by the DOI.
