+++ man/mx_api_content.Rd
@@ -18,18 +18,22 @@ mx_api_content(
 
 \item{to_date}{Latest date of interest. Defaults to current date.}
 
-\item{clean}{Logical, indicating whether to clean the data returned for use
-with other mx_* functions. Default is TRUE.}
+\item{clean}{Logical, defaulting to TRUE, indicating whether to clean the
+data returned by the API. If TRUE, variables containing absolute paths to
+the preprints web-page ("link_page") and PDF ("link_pdf") are generated
+from the "server", "DOI", and "version" variables returned by the API. The
+"title", "abstract" and "authors" variables are converted to title case.
+Finally, the "type" and "server" variables are dropped.}
 
 \item{server}{Specify the server you wish to use: "medrxiv" (default) or
 "biorxiv"}
 
 \item{include_info}{Logical, indicating whether to include variables
-containing information returned by the API (e.g. cursor number, total count
-of papers, etc). Default is FALSE.}
+containing information returned by the API (e.g. API status, cursor number,
+total count of papers, etc). Default is FALSE.}
 }
 \value{
-Dataframe with 1 record per row
+Tibble with 1 record per row
 }
 \description{
 Provides programmatic access to all preprints available through
