+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
 pkgdown: 1.5.1.9000
-pkgdown_sha: eae56f08694abebf93cdfc0dd8e9ede06d8c815f
+pkgdown_sha: 97b2f9b1d2fb9f5f4e82b6d661197c4b4b49368b
 articles: []
-last_built: 2020-08-14T16:05Z
+last_built: 2020-08-24T13:35Z
 
