### 2020-08-24 cc2274f
Update medrxiv-api.Rmd
Change so that it presents not cleaned versiom
