### 2020-08-24 #380 @mcguinlu
Thanks @njahn82. Just to note as well that I recently moved the snapshot functionality from relying on my local Task Scheduler to working from [GitHub Actions](https://github.com/mcguinlu/medrxivr-data), so it should now be a lot more robust (in the past, if my local PC experienced network issues, the snapshot would not be taken).

In response to your comments:

**I wonder if the returned data frames from the mx_api_*() family could be also represented as tibbles?**
  * The package now returns tibbles across the board. I had never really understood the difference, but after a bit of research, I do prefer the printing defaults for `tibble` objects.

**The package does a good job in parsing and cleaning preprint metadata. Unfortunately, I cannot find documentation or an example showcasing what is actually returned. Can you provide one reproducible example in the README and/or extend the documentation in the function docs?**
  * Hoping I understood this ask correctly, there is now a [section in the README](https://github.com/mcguinlu/medrxivr#accessing-the-raw-api-data) that desribes how to access the raw, uncleaned API data using the `mx_api_*()` functions, which also points to a section in the [API article](https://mcguinlu.github.io/medrxivr/articles/medrxiv-api.html#accessing-the-raw-api-data) on the `pkgdown` website that provides more detail and an example of the uncleaned output. In addition, a clearer description of what the cleaning process entails has been included in the documention of the `mx_api_*()` functions (e.g. [here](https://mcguinlu.github.io/medrxivr/reference/mx_api_content.html))

**In the function docs of mx_export(), it says Dataframe returned by mx_search(), but I realised that also data obtained from the mx_api_ family can be exported as bib file using mx_export().**
  * Thanks for this - I have updated the [docs for the `mx_export()` function](https://mcguinlu.github.io/medrxivr/reference/mx_export.html) to read `@param data Dataframe returned by mx_search() or mx_api_*() functions`

@maurolepore, I have checked that these changes don't throw any new errors and that `goodpractice` doesn't recommend any changes. I've also re-run `styler`/`spelling` functions and commited any modications.

Hoping we are nearly there!
