+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.7.3
 pkgdown: 1.5.1.9000
 pkgdown_sha: 97b2f9b1d2fb9f5f4e82b6d661197c4b4b49368b
 articles: []
-last_built: 2020-08-24T13:35Z
+last_built: 2020-08-24T13:50Z
 
