### 2020-08-28 #380 @maurolepore
Thanks @njahn82 and  @mcguinlu,

@njahn82, once again, please consider @mcguinlu 's changes and respond with either your approval or further suggestions for improvement.
