### 2020-09-05 #380 @maurolepore
Approved! Thanks @mcguinlu  for submitting and @tts and @njahn82 for your reviews! :smile:

- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  Soon you will be invited to a team that should allow you to do so.  You'll be made admin once you do.

From https://github.com/ropensci/software-review/issues/380#issue-625738088 I see you wish to automatically submit to the Journal of Open Source Software? If so:

- [ ] Submit to JOSS at <https://joss.theoj.org/papers/new>, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`
