### 2020-09-08 #380 @mcguinlu
Hi @annakrystalli have transferred across now. @maurolepore thanks for the checklist - I will work through it over the coming day. And finally, just flagging to @stefaniebutland that I would be interested in producing a blog post for this package!

Thanks again to @tts and @njahn82 for reviewing, and @maurolepore for herding us all through the process!
