### 2020-09-08 #380 @annakrystalli
Thanks @mcguinlu ! Full admin rights now returned 👍
