### 2020-09-08 #380 @mcguinlu
> To-dos:
>
> * [x]  Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  Soon you will be invited to a team that should allow you to do so.  You'll be made admin once you do.
> * [x]  Fix all links to the GitHub repo to point to the repo under the ropensci organization.
> * [x]  If you already had a `pkgdown` website **and are ok relying only on [rOpenSci central docs building and branding](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs)**,
>
>   * deactivate the automatic deployment you might have set up
>   * remove styling tweaks from your pkgdown config but keep that config file
>   * replace the whole current `pkgdown` website with a [redirecting page](https://devguide.ropensci.org/#redirect)
>   * replace your package docs URL with `https://docs.ropensci.org/package_name`
>   * In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
> * [x]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`.
> * [x]  We're starting to roll out software metadata files to all ropensci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
>
> From [#380 (comment)](https://github.com/ropensci/software-review/issues/380#issue-625738088) I see you wish to automatically submit to the Journal of Open Source Software? If so:
>
> * [x]  Activate Zenodo watching the repo
> * [x]  Tag and create a release so as to create a Zenodo version and DOI
> * [x]  Submit to JOSS at https://joss.theoj.org/papers/new, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`

Okay, I've completed all the steps now @maurolepore! Re: the JOSS review, please see @danielskatz's comment above.

The one thing I wasn't clear on was how to replace the old `pkgdown` website with a redirecting page - seeing as the repo has been transferred across, the old `pkgdown` website on GitHub Pages ([https://mcguinlu.github.io/medrxivr/index.html](https://mcguinlu.github.io/medrxivr/index.html)) no longer exists (I think?!), so I wasn't clear on how to set the redirect.
