### 2020-09-08 #380 @maurolepore
@mcguinlu,

RE:
> The one thing I wasn't clear on was how to replace the old pkgdown website with a redirecting page - seeing as the repo has been transferred across, the old pkgdown website on GitHub Pages (https://mcguinlu.github.io/medrxivr/index.html) no longer exists (I think?!), so I wasn't clear on how to set the redirect.

I'm sorry this isn't clear for you or me. But as you say, the working website seems correct. I see no reason to worry.

Here are a few more comments from section 8.1.4 of <https://devguide.ropensci.org/>:

* If you intend to submit to CRAN, see [CRAN gotchas](https://devguide.ropensci.org/building.html#crangotchas). I"m happy to provide support through this process. Let me know.

Please check these boxes to confirm you've done the following last steps:

- [ ] Add a CodeMeta file by running codemetar::write_codemeta() (codemetar GitHub repo)
- [ ] Change any needed links, such those for CI badges
- [ ] Re-activate CI services
    * For Travis, activating the project in the ropensci account should be sufficient
    * For AppVeyor, tell the author to update the GitHub link in their badge, but do not transfer the project: AppVeyor projects should remain under the authors’ account. The badge is [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname).
    * For Codecov, the webhook may need to be reset by the author.

- [ ] If authors maintain a gitbook that is at least partly about their package, contact an rOpenSci staff member so they might contact the authors about transfer to the ropensci-books GitHub organisation.

- [ ] Add a “peer-reviewed” topic to the repo (it seems I'm the one supposed to do this but I apparently lack the privileges to access the "topics" settings -- see if you can or let me know).

--

Ping me when this is done and I'll then close this issue.

Thanks!
