### 2020-09-08 #380 @maurolepore
@mcguinlu, I see you already mentioned Stephanie Butland above. To comply with https://devguide.ropensci.org/editorguide.html#after-review, I here I also mention ropensci/blog-editors for follow-up about your willingness to write a blog post or tech note.

Finally, please see https://devguide.ropensci.org/editorguide.html#package-promotion
