### 2020-09-09 #380 @maurolepore
@mcguinlu , thanks and congratulations! To the best of my knowledge, this completes the review process so I'll close now.

--

You may already know this. To prepare packages for CRAN, [`usethis::use_release_issue()`](https://usethis.r-lib.org/reference/use_release_issue.html) is useful. And [here](https://github.com/2DegreesInvesting/resources/issues/201) are some aspects of the workflow that might help (including a link to some [tweaks](https://github.com/2DegreesInvesting/resources/issues/198)). Feel free to reach out with questions.

Are you in rOpenSci's Slack workspace? If not, I recommend you find someone who can add you. I have found friendly advice there that I wouldn't find anywhere else.
