### 2020-09-09 #380 @stefaniebutland
Hello @mcguinlu. We'd love to have a post about medrxivr.

Our [Blog Guide](https://blogguide.ropensci.org/) has most of the information you should need, with both content and technical advice.

Let me know when you'd like to submit a draft and I can suggest a publication date.
