### 2020-09-09 #380 @stefaniebutland
@mcguinlu Also let me know if you'd like a new invitation to rOpenSci Slack
