### 2020-09-10 #380 @mcguinlu
@stefaniebutland a new invite would be great! I thought I had activated the first one correctly but apparently not (I am still getting to grips with Slack) 🤦‍♂️ and happy to continue chatting about this there.
